[REFAC] project file architecture
