+++ lib/src/modules/base/decoration_base.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/material.dart';
+
+abstract class DecorationBase {
+  final Color color;
+  final double strokeWidth;
+
+  const DecorationBase({required this.color, required this.strokeWidth});
+}
