+++ lib/src/modules/base/styles.dart
@@ -1,0 +1,4 @@
+export '../box/enums/box_style.dart';
+export '../circle/enums/circle_style.dart';
+export '../highlight/enums/highlight_style.dart';
+export '../underline/enums/underline_style.dart';
