+++ lib/src/modules/box/classes/box_size.dart
