+++ lib/src/modules/circle/classes/circle_size.dart
@@ -1,0 +1,13 @@
+class CircleSize {
+  final double textWidth;
+  final double textHeight;
+  final double horizontalRadius;
+  final double verticalRadius;
+
+  CircleSize({
+    required this.textWidth,
+    required this.textHeight,
+    required this.horizontalRadius,
+    required this.verticalRadius,
+  });
+}
