+++ lib/src/modules/highlight/painter/highlight_painter.dart
