+++ lib/src/modules/underline/classes/horizontal_offset.dart
@@ -1,0 +1,5 @@
+class HorizontalOffset {
+  final double left;
+  final double right;
+  const HorizontalOffset(this.left, this.right);
+}
