+++ lib/src/widgets/text_draw_decorator.dart
@@ -1,91 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_text_draw_decorator/flutter_text_draw_decorator.dart';
-import 'package:flutter_text_draw_decorator/src/painter/box/box_painter.dart';
-import 'package:flutter_text_draw_decorator/src/painter/box/wavy_box_painter.dart';
-import 'package:flutter_text_draw_decorator/src/painter/highlight/highlight_painter.dart';
-import 'package:flutter_text_draw_decorator/src/painter/underline/underline_painter.dart';
-
-// TODO: Cleanup optionals, maybe use classes instead of enum?
-class CircleDecoration extends DecorationBase {
-  const CircleDecoration({required super.color, required super.strokeWidth});
-
-  factory CircleDecoration.standard() {
-    return const CircleDecoration(color: Colors.orange, strokeWidth: 1);
-  }
-}
-
-enum CircleStyle {
-  openCircled,
-  closedCircled;
-  //...
-
-  CustomPainter getPainter(Text text, {CircleDecoration decoration = const CircleDecoration(color: Colors.orange, strokeWidth: 1)}) {
-    switch (this) {
-      case CircleStyle.openCircled:
-        return OpenCirclePainter(text: text.data ?? '', textStyle: text.style ?? const TextStyle(), decoration: decoration);
-      case CircleStyle.closedCircled:
-        return ClosedCirclePainter(text: text.data ?? '', textStyle: text.style ?? const TextStyle(), decoration: decoration);
-      //...
-    }
-  }
-}
-
-enum UnderlineStyle {
-  horizontal,
-  curved;
-  //...
-
-  CustomPainter getPainter(Text text, Color color, double strokeWidth) {
-    switch (this) {
-      case UnderlineStyle.horizontal:
-        return HorizontalUnderlinePainter(color: color, strokeWidth: strokeWidth);
-      case UnderlineStyle.curved:
-        return CurvedUnderlinePainter(textStyle: text.style ?? const TextStyle(), color: color, text: text.data ?? '', strokeWidth: strokeWidth);
-      //...
-    }
-  }
-}
-
-enum BoxStyle {
-  rounded,
-  bubble,
-  wavy;
-
-  CustomPainter getPainter(Text text, double borderRadius, double strokeWidth) {
-    switch (this) {
-      case BoxStyle.rounded:
-        return RoundedBoxPainter(text: text, borderRadius: borderRadius, strokeWidth: strokeWidth);
-      case BoxStyle.bubble:
-        return BubbleBoxPainter(
-          text: text,
-          padding: 4,
-          bubbleColor: Colors.orange,
-          tip: const BubbleBoxTip(
-            position: TipPosition.left,
-            orientation: TipOrientation.left,
-          ),
-        );
-      case BoxStyle.wavy:
-        return WavyBoxPainter(text: text, borderColor: Colors.black);
-    }
-  }
-}
-
-enum HighlightStyle {
-  textmarker;
-
-  CustomPainter getPainter(Text text, Color? color, double? strokeWidth) {
-    switch (this) {
-      case HighlightStyle.textmarker:
-        return HighlightPainter(
-          text: text.data ?? '',
-          color: color ?? Colors.yellow,
-          textStyle: text.style ?? const TextStyle(),
-          strokeWidth: strokeWidth,
-        );
-    }
-  }
-}
+import 'package:flutter_text_draw_decorator/src/modules/box/enums/box_style.dart';
+import 'package:flutter_text_draw_decorator/src/modules/circle/enums/circle_style.dart';
+import 'package:flutter_text_draw_decorator/src/modules/highlight/enums/highlight_style.dart';
+import 'package:flutter_text_draw_decorator/src/modules/underline/enums/underline_style.dart';
+import 'package:flutter_text_draw_decorator/src/modules/circle/decorations/circle_decoration.dart';
 
 class TextDrawDecorator extends StatelessWidget {
   final Text text;
@@ -96,6 +14,14 @@ class TextDrawDecorator extends StatelessWidget {
     required this.painter,
     super.key,
   });
+
+  @override
+  Widget build(BuildContext context) {
+    return CustomPaint(
+      painter: painter,
+      child: text,
+    );
+  }
 
   factory TextDrawDecorator.circled({
     required Text text,
@@ -139,13 +65,5 @@ class TextDrawDecorator extends StatelessWidget {
     double? strokeWidth,
   }) {
     return TextDrawDecorator(text: text, painter: style.getPainter(text, color, strokeWidth));
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return CustomPaint(
-      painter: painter,
-      child: text,
-    );
   }
 }
